+++ lib/boost.dart
@@ -6,6 +6,7 @@ export 'src/notifier.dart';
 export 'src/strings.dart';
 export 'src/structures.dart';
 export 'src/types.dart';
+export 'src/math.dart';
 export 'src/concurrency/canceled_exception.dart';
 export 'src/concurrency/cancellation_token.dart';
 export 'src/concurrency/guard.dart';